pyproject.toml
[tool.poetry]
name = "postman-docs-ai"
version = "0.1.0"
description = ""
authors = ["John Walz <johnwalz97@gmail.com>"]
readme = "README.md"
packages = [{include = "postman_docs_ai"}]

[tool.poetry.dependencies]
python = "^3.11,<3.12"
openai = {extras = ["datalib"], version = "^0.27.4"}
python-dotenv = "^1.0.0"
qdrant-client = "^1.1.4"
uuid = "^1.30"
fastapi = {extras = ["all"], version = "^0.95.1"}
uvicorn = "^0.21.1"
tiktoken = "^0.3.3"


[tool.poetry.group.dev.dependencies]
pyyaml = "^6.0"
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.261"
marko = "^1.3.0"
tqdm = "^4.65.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
